move Core/CommonTypedefs.h to PlatformAgnostic
