+++ chakracore-cxx/lib/Common/CommonBasic.h
@@ -50,7 +50,7 @@ using std::type_info;
 #include "Core/CommonMinMax.h"
 
 // === Core Header Files ===
-#include "Core/CommonTypedefs.h"
+#include "PlatformAgnostic/CommonTypedefs.h"
 #include "Core/Api.h"
 #include "Core/CriticalSection.h"
 #include "Core/Assertions.h"
