+++ chakracore-cxx/lib/Common/PlatformAgnostic/CommonTypedefs.h
