+++ chakracore-cxx/lib/Runtime/PlatformAgnostic/UnicodeText.h
@@ -4,7 +4,7 @@
 //-------------------------------------------------------------------------------------------------------
 #pragma once
 
-#include "Core/CommonTypedefs.h"
+#include "PlatformAgnostic/CommonTypedefs.h"
 #include "ChakraICU.h"
 #include "sal.h"
